sveltekit: add context provider component for mocking sveltekit contexts
