+++ code/frameworks/sveltekit/package.json
@@ -32,6 +32,7 @@
       "types": "./dist/index.d.ts",
       "default": "./dist/index.js"
     },
+    "./internal/MockProvider.svelte": "./static/MockProvider.svelte",
     "./internal/mocks/app/forms": "./dist/mocks/app/forms.js",
     "./internal/mocks/app/navigation": "./dist/mocks/app/navigation.js",
     "./internal/mocks/app/stores": "./dist/mocks/app/stores.js",
@@ -46,6 +47,7 @@
   },
   "files": [
     "dist/**/*",
+    "static/**/*",
     "template/**/*",
     "README.md",
     "*.js",
