+++ code/frameworks/sveltekit/tsconfig.json
@@ -4,7 +4,6 @@
     "paths": {
       "storybook/internal/*": ["../../lib/cli/core/*"]
     },
-    "rootDir": "./src"
   },
   "extends": "../../tsconfig.json",
   "include": ["src/**/*"]
